+++ client/cypress.config.ts
@@ -1,15 +1,19 @@
 import { defineConfig } from 'cypress'
+import {
+  getPdfContent,
+  setupMockSmtpServer,
+  configureDownloadDirectory,
+} from './cypress/plugins'
 
 export default defineConfig({
   viewportWidth: 1000,
   viewportHeight: 1000,
   defaultCommandTimeout: 10000,
   e2e: {
-    // We've imported your old cypress plugins here.
-    // You may want to clean this up later by importing these.
-    setupNodeEvents(on, config) {
-      // eslint-disable-next-line global-require
-      return require('./cypress/plugins/index.js')(on, config)
+    setupNodeEvents(on, _config) {
+      on('task', { getPdfContent })
+      configureDownloadDirectory(on)
+      setupMockSmtpServer(on)
     },
     baseUrl: 'http://localhost:3000',
     specPattern: 'cypress/end-to-end/**/*.cy.{js,jsx,ts,tsx}',
